Refactor :: 프로필 생성 관리 기능 변경
CreateProfileService를 ManageProfileService로 통합하여 기존 및 신규 프로필 관리 기능을 추가했습니다. 연결된 인터페이스 이름 및 호출부를 반영했습니다.
