+++ src/main/kotlin/com/seugi/api/domain/profile/application/port/in/ManageProfileUseCase.kt
@@ -3,8 +3,8 @@ package com.seugi.api.domain.profile.application.port.`in`
 import com.seugi.api.domain.workspace.domain.enums.WorkspaceRole
 
 
-interface CreateProfileUseCase {
+interface ManageProfileUseCase {
 
-    fun createProfile(memberId: Long, workspaceId: String, permission: WorkspaceRole)
+    fun manageProfile(memberId: Long, workspaceId: String, permission: WorkspaceRole)
 
 }
